project2_implementation/pyproject.toml
[tool.poetry]
name = "Harshad Bhandwaldar"
version = "1.0.0"
description = "Data Science project"
authors = ["Harshad Bhandwaldar"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.85.1"
uvicorn = "^0.18.2"
pydantic = "^1.9.1"
MyApplication = "^0.1.0"
isort = "^5.10.1"
black = "^22.6.0"
flake8 = "^4.0.1"
pandas = "^1.5.1"
mysql-connector-python = "^8.0.31"
matplotlib = "^3.6.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
pytest = "^7.0.1"
mypy = "^0.931"

[tool.isort]
profile = "black"

[tool.mypy]
# unfortunately cannot inject dependencies into super-linter
# so we have to ignore the imports as they will not have typehints
ignore_missing_imports = true
warn_unused_configs = true
disallow_any_generics = true
# disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
# disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
experimental-string-processing = true